# src/main/java/com/wirvsvirus/backend/model/Questionnaire.java
package com.wirvsvirus.backend.model;

import com.fasterxml.jackson.annotation.JsonIgnore;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import org.hibernate.annotations.Type;
import org.springframework.hateoas.RepresentationModel;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.persistence.Transient;
import java.sql.Timestamp;
import java.util.List;

@EqualsAndHashCode(callSuper = true)
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name = "questionnaires")
public class Questionnaire extends RepresentationModel<Questionnaire> {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long questionnaireId;
    private Long userId;
    private Timestamp date;
    @Type(type = "org.hibernate.type.NumericBooleanType")
    @JsonProperty("is_completed")
    private boolean isCompleted;

    private String version;
    @JsonIgnore
    private Long start;

    @Transient
    private List<Question> questions;
}